Disable Dependabot
We don't have production dependencies, only dev dependencies for running a
web server and doing linting. The update to ESLint 9 is a significant
change and until we have done this, we don't want to litter this repo
with failing pull requests for ESLint or its plugins.

Ticket: https://phabricator.wikimedia.org/T387567